add authentication

+++ backend/controllers/internship.js
@@ -2,14 +2,19 @@ import { pool } from '../config/database.js'
 
 
 const createInternship = async (req, res) => {
+    if (!req.isAuthenticated()) {
+        return res.status(401).json({ error: 'Not authenticated' });
+    }
+
     try {
         const { company, position, content, url, category_id, status_id } = req.body
+        const user_id = req.user.id
 
         const results = await pool.query(`
-            INSERT INTO internships (company, position, content, url, category_id, status_id)
-            VALUES ($1, $2, $3, $4, $5, $6)
+            INSERT INTO internships (company, position, content, url, category_id, status_id, user_id)
+            VALUES ($1, $2, $3, $4, $5, $6, $7)
             RETURNING *`,
-            [company, position, content, url, category_id, status_id]
+            [company, position, content, url, category_id, status_id, user_id]
         )
         res.status(201).json(results.rows[0])
     }
@@ -19,8 +24,14 @@ const createInternship = async (req, res) => {
 }
 
 const getAllInternships = async (req, res) => {
+    if (!req.isAuthenticated()) {
+        return res.status(401).json({ error: 'Not authenticated' });
+    }
+
     try {
-        const results = await pool.query('SELECT * FROM internships ORDER BY created_at DESC')
+        const user_id = req.user.id
+
+        const results = await pool.query('SELECT * FROM internships WHERE user_id = $1 ORDER BY created_at DESC', [user_id])
         res.status(200).json(results.rows)
     }
     catch (error) {
@@ -29,9 +40,14 @@ const getAllInternships = async (req, res) => {
 }
 
 const getInternshipById = async (req, res) => {
+    if (!req.isAuthenticated()) {
+        return res.status(401).json({ error: 'Not authenticated' });
+    }
+
     try {
         const id = req.params.id
-        const results = await pool.query('SELECT * FROM internships WHERE id = $1', [id])
+        const user_id = req.user.id
+        const results = await pool.query('SELECT * FROM internships WHERE id = $1 AND user_id = $2', [id, user_id])
         res.status(200).json(results.rows[0])
     }
     catch (error) {
@@ -40,15 +56,20 @@ const getInternshipById = async (req, res) => {
 }
 
 const updateInternship = async (req, res) => {
+    if (!req.isAuthenticated()) {
+        return res.status(401).json({ error: 'Not authenticated' });
+    }
+
     try {
         const id = req.params.id
+        const user_id = req.user.id
         const { company, position, content, url, category_id, status_id } = req.body
         const { rows: updatedInternships } = await pool.query(
             `UPDATE internships
             SET company = $1, position = $2, content = $3, url = $4, category_id = $5, status_id = $6
-            WHERE id = $7
+            WHERE id = $7 AND user_id = $8
             RETURNING *`,
-            [company, position, content, url, category_id, status_id, id]
+            [company, position, content, url, category_id, status_id, id, user_id]
         )
         res.status(200).json(updatedInternships[0])
     }
@@ -58,9 +79,14 @@ const updateInternship = async (req, res) => {
 }
 
 const deleteInternship = async (req, res) => {
+    if (!req.isAuthenticated()) {
+        return res.status(401).json({ error: 'Not authenticated' });
+    }
+
     try {
+        const user_id = req.user.id
         const id = req.params.id
-        const results = await pool.query('DELETE FROM internships WHERE id = $1 RETURNING *', [id])
+        const results = await pool.query('DELETE FROM internships WHERE id = $1  AND user_id = $2 RETURNING * ', [id, user_id])
         res.status(200).json(results.rows[0])
     }
     catch (error) {
