+++ backend/controllers/project.js
@@ -3,13 +3,14 @@ import { pool } from '../config/database.js'
 
 const createProject = async (req, res) => {
   try {
+    const user_id = req.user.id
     const { title, content, url, category_id, status_id } = req.body
 
     const results = await pool.query(`
-        INSERT INTO projects (title, content, url, category_id, status_id)
-        VALUES ($1, $2, $3, $4, $5)
+        INSERT INTO projects (title, content, url, category_id, status_id, user_id)
+        VALUES ($1, $2, $3, $4, $5, $6)
         RETURNING *`,
-      [title, content, url, category_id, status_id]
+      [title, content, url, category_id, status_id, user_id]
     )
     res.status(201).json(results.rows[0])
   }
@@ -20,7 +21,8 @@ const createProject = async (req, res) => {
 
 const getAllProjects = async (req, res) => {
   try {
-    const results = await pool.query('SELECT * FROM projects ORDER BY updated_at DESC')
+    const user_id = req.user.id
+    const results = await pool.query('SELECT * FROM projects WHERE user_id = $1 ORDER BY updated_at DESC', [user_id])
     res.status(200).json(results.rows)
   }
   catch (error) {
@@ -30,6 +32,7 @@ const getAllProjects = async (req, res) => {
 
 const getProjectById = async (req, res) => {
   try {
+    const user_id = req.user.id
     const id = req.params.id
     const results = await pool.query('SELECT * FROM projects WHERE id = $1', [id])
     res.status(200).json(results.rows[0])
@@ -41,14 +44,15 @@ const getProjectById = async (req, res) => {
 
 const updateProject = async (req, res) => {
   try {
+    const user_id = req.user.id
     const id = req.params.id
     const { title, content, url, category_id, status_id } = req.body
     const results = await pool.query(
       `UPDATE projects
       SET title = $1, content = $2, url = $3, category_id = $4, status_id = $5
-      WHERE id = $6
+      WHERE id = $6 AND user_id = $7
       RETURNING *`,
-      [title, content, url, category_id, status_id, id]
+      [title, content, url, category_id, status_id, id, user_id]
     )
     res.status(200).json(results.rows[0])
   }
@@ -59,8 +63,9 @@ const updateProject = async (req, res) => {
 
 const deleteProject = async (req, res) => {
   try {
+    const user_id = req.user.id
     const id = req.params.id
-    const results = await pool.query('DELETE FROM projects WHERE id = $1 RETURNING *', [id])
+    const results = await pool.query('DELETE FROM projects WHERE id = $1 AND user_id = $2 RETURNING *', [id, user_id])
     res.status(200).json(results.rows[0])
   }
   catch (error) {
